Vendor Complete

+++ src/db/Models/BrandCategories.js
@@ -3,17 +3,17 @@ module.exports = (sequelize, DataTypes) => {
 
     const BrandCategories = sequelize.define("brand_category", {
         id: {
-            type: DataTypes.BIGINT,
+            type: DataTypes.INTEGER,
             primaryKey: true,
             allowNull: false,
             autoIncrement: true
         },
         brand_id: {
-            type: DataTypes.BIGINT,
+            type: DataTypes.INTEGER,
             allowNull: false
         },
         cat_id: {
-            type: DataTypes.BIGINT,
+            type: DataTypes.INTEGER,
             allowNull: false
         },
         tenant_id: {
